fix(tap-to-walk): calibrate default scale for Pokémon Go, switch to slider
Default scale 0.23 m/px (measured from two pokéstops ~40.5 m apart spanning
~173 px at full zoom-out). Range narrowed to 0.01–1.0 (Double) replacing the
old 1–50 Int field. Text input replaced with a continuous slider; description
now reminds users to zoom fully out for best accuracy.

+++ core/data/src/main/kotlin/com/locationjoystick/core/data/SettingsRepository.kt
@@ -214,7 +214,7 @@ class SettingsRepository
 
         fun getTapToWalkOverlayEnabled(): Flow<Boolean> = dataSource.getTapToWalkOverlayEnabled()
 
-        fun getTapToWalkScaleMpx(): Flow<Int> = dataSource.getTapToWalkScaleMpx()
+        fun getTapToWalkScaleMpx(): Flow<Double> = dataSource.getTapToWalkScaleMpx()
 
         fun getCompassTrackingEnabled(): Flow<Boolean> = dataSource.getCompassTrackingEnabled()
