+++ feature/settings/impl/src/main/kotlin/com/locationjoystick/feature/settings/impl/SettingsUiState.kt
@@ -36,7 +36,7 @@ data class SettingsUiState(
     val enabledMapFeatures: Set<AppFeature> = AppFeature.DEFAULT_MAP_ENABLED,
     val floatingMapQuickWalk: Boolean = false,
     val tapToWalkOverlayEnabled: Boolean = false,
-    val tapToWalkScaleMpx: Int = AppConstants.TapToWalkConstants.DEFAULT_SCALE_MPX,
+    val tapToWalkScaleMpx: Double = AppConstants.TapToWalkConstants.DEFAULT_SCALE_MPX,
     val compassTrackingEnabled: Boolean = false,
     val isCompassServiceGranted: Boolean = false,
     val compassRegionCxPct: Float = AppConstants.CompassTrackingConstants.DEFAULT_REGION_CX_PCT,
