+++ feature/settings/impl/src/test/kotlin/com/locationjoystick/feature/settings/impl/SettingsViewModelSaveTest.kt
@@ -560,7 +560,7 @@ internal class SaveTestPreferencesDataSource : PreferencesDataSource {
 
     override fun getTapToWalkOverlayEnabled(): Flow<Boolean> = flowOf(false)
 
-    override fun getTapToWalkScaleMpx(): Flow<Int> = flowOf(AppConstants.TapToWalkConstants.DEFAULT_SCALE_MPX)
+    override fun getTapToWalkScaleMpx(): Flow<Double> = flowOf(AppConstants.TapToWalkConstants.DEFAULT_SCALE_MPX)
 
     override fun getCompassTrackingEnabled(): Flow<Boolean> = flowOf(false)
